Schedule Halide functions by internal name
The autotuner now builds a map of Halide-internal name to function object
which it uses to access functions by Halide name for scheduling. This
avoids the problem of requiring names used by the tuner to all be visible
as in-scope C objects pointing to the corresponding Func at the
AUTOTUNE_HOOK point.