Add fourscore flag to fm2 header
While the fourscore key is not shown as required in the fm2 format
documentation, different versions of the emulator assume different
values if it isn't present, causing input to be parsed differently.
From OpenTuner's perspective (assuming 0), some input is being skipped
by the emulator (assuming 1).  OpenTuner still learns just fine but
playing back the concatenation of all the inputs (as in the new-bests
movie) does not work.  Note that it's not strictly every other frame
being skipped, so the obvious solution of manual decimation to convert
from assumes-1 to assumes-0 input files does not work.