--xvfb-delay option
The demo works fine with 0 on my local VM, but lanka needs 1, and other
systems may need more.  Tthe xvfb-run default is apparently 3, though
manual inspection of the script on my machine shows it receives a signal
from xvfb when xvfb is ready for connections, and so doesn't wait at
all.