Mario: instantiate defaults (i.e., make defaults actually work)
argparse doesn't apply the action when using the default argument, but
it will call the type if the argument is a string
(http://stackoverflow.com/a/21588198/3614835).

We could also have fixed this by passing instances as the defaults
(instead of class names), but that requires having parsed their
definitions already, and it's OpenTuner style to declare options near
the top of the file.  By using type, we delay the instantiation until
after the representation/duration classes have been parsed.