Mario example updates (#96)
* Mario: shebang uses env python2

GitHub issue #57 proposes to do this globally and add a git hook to
enforce it for the future.  I'm doing this one separately because it's
immediately blocking me.

* Mario: tuning-run parameter is an int

* Mario: instantiate defaults (i.e., make defaults actually work)

argparse doesn't apply the action when using the default argument, but
it will call the type if the argument is a string
(http://stackoverflow.com/a/21588198/3614835).

We could also have fixed this by passing instances as the defaults
(instead of class names), but that requires having parsed their
definitions already, and it's OpenTuner style to declare options near
the top of the file.  By using type, we delay the instantiation until
after the representation/duration classes have been parsed.

* Mario: remove --fceux-path option

The user should just set PATH manually before running the tuner.

* Mario: centralize Popen error handling

* Mario: fail-fast PATH checks

* Mario: check sqlite3 return code

This fixes jansel/opentuner#92.

* Mario: tell xvfb-run to print its errors

* Mario: manage Xvfb display numbers

xvfb-run's --auto-servernum is vulnerable to race conditions (see
http://stackoverflow.com/a/30336424/3614835), so we have to manage them
explicitly.

* Mario: implement save_final_config

* Mario: update README.md